tag refactoring

+++ src/main/java/it/linkshare/service/TagService.java
@@ -1,15 +1,17 @@
 package it.linkshare.service;
 
-import it.linkshare.controller.dto.Tag;
-import it.linkshare.controller.dto.TagCreationRequest;
+import it.linkshare.dao.TagDAO;
+import it.linkshare.dto.TagRequestDTO;
+import it.linkshare.dto.TagResponseDTO;
+import it.linkshare.mapper.TagMapper;
 import it.linkshare.repository.TagRepository;
-import it.linkshare.repository.entity.TagEntity;
-import it.linkshare.service.mapper.TagMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class TagService {
@@ -20,35 +22,32 @@ public class TagService {
         this.tagRepository = tagRepository;
     }
 
-    public Tag getTagById(Long id) {
-        return tagRepository.findFirstById(id)
-                .map(TagMapper::mapToRest)
+    public TagResponseDTO getTagById(Long id) {
+        return tagRepository.findById(id)
+                .map(TagMapper::mapToResponseDTO)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
 
-    public List<TagEntity> getAllTag() {
-        return tagRepository.findAll();
+    public List<TagResponseDTO> getAllTags() {
+        return tagRepository.findAll().stream().map(TagMapper::mapToResponseDTO).collect(Collectors.toList());
     }
 
-    public Tag addNewTag(TagCreationRequest request) {
-        return TagMapper.mapToRest(tagRepository.save(TagMapper.mapToEntity(request)));
+    public TagResponseDTO addNewTag(TagRequestDTO tagRequestDTO) {
+        return TagMapper.mapToResponseDTO(tagRepository.save(TagMapper.mapToDAO(tagRequestDTO)));
     }
 
-    public TagEntity updateTag(TagEntity tag, Long id) {
-        return tagRepository.findById(id)
-                .map(t -> {
-                    t.setName(tag.getName());
-                    t.setNsfw(tag.getNsfw());
-                    return tagRepository.save(t);
-                })
-                .orElseGet(() -> {
-                    tag.setId(id);
-                    return tagRepository.save(tag);
-                });
+    public TagResponseDTO updateTag(TagRequestDTO tagRequestDTO, Long id) {
+        TagDAO tagDAO = tagRepository.findById(id)
+                .map(tag -> {
+                    tag.setName(tagRequestDTO.getName());
+                    tag.setNsfw(tagRequestDTO.getNsfw());
+                    return tag;})
+                .orElse(TagMapper.mapToDAO(tagRequestDTO, id));
+        return TagMapper.mapToResponseDTO(tagRepository.save(tagDAO));
     }
 
     public void deleteTag(Long id) {
-        tagRepository.deleteById(id);
+        if(tagRepository.existsById(id))
+            tagRepository.deleteById(id);
     }
-
 }
